extract directives into src/directives
