+++ src/components/Data/Collections/Watch.vue
@@ -129,7 +129,7 @@
   import Headline from '../../Materialize/Headline'
   import JsonFormatter from '../../../directives/json-formatter.directive'
   import ScrollGlue from '../../../directives/scroll-glue.directive'
-  import jQueryCollapsible from '../../Materialize/collapsible'
+  import jQueryCollapsible from '../../../directives/collapsible.directive'
   import Notification from '../Realtime/Notification'
   import SubscriptionControls from '../Realtime/SubscriptionControls'
   import CollectionDropdown from '../Collections/Dropdown'
