+++ src/directives/collapsible.directive.js
